refactor: common 패키지 내 entity 구조 정리
