+++ src/main/java/com/example/sns_feed/user/entity/User.java
@@ -1,6 +1,6 @@
 package com.example.sns_feed.user.entity;
 
-import com.example.sns_feed.common.BaseEntity;
+import com.example.sns_feed.common.entity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.Getter;
